COSMO/JuMP extension: ex8 runs
