+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["G Jake Gebbie <ggebbie@whoi.edu>"]
 version = "0.2.17"
 
 [deps]
+COSMO = "1e616198-aa4e-51ec-90a2-23f7fbd31d8d"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -12,6 +13,7 @@ GeoPythonPlot = "1e05a8e1-7dec-4f9e-9d3d-7df52321841b"
 GibbsSeaWater = "9a22fb26-0b63-4589-b28e-8f9d0b5c3d05"
 GoogleDrive = "91feb7a0-3508-11ea-1e8e-afea2c1c9a19"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
